social1/requirements.txt
asgiref==3.5.2
Django==4.1.1
django-crispy-forms==1.14.0
docopt==0.6.2
greenlet==1.1.3
Pillow==9.2.0
records==0.5.2
SQLAlchemy==1.4.41
sqlparse==0.4.3
tablib==3.2.1
tzdata==2022.4
whitenoise==6.2.0
django-allauth==0.59.0
six==1.17.0
<<<<<<< HEAD
=======
pymysql==1.1.2
django-extensions==3.2.3
>>>>>>> origin/izmene
